// src/types/GameStatus.ts
export enum PLAYER_STATUS {
  GAME_START_WAIT = 'Game started! Wait your opponent.',
  GAME_START_TURN = 'Game started! Your turn:',
  DRAW = 'Draw!',
  WIN = 'You win!',
  LOST = 'You lost!',
  WAIT_OPPONENT = 'Wait your opponent.',
  YOUR_TURN = 'Your turn:'
}

export enum GAME_STATUS {
  WIN_LOST = 'win-lost',
  DRAW = 'draw',
  PLAYING = 'playing' 
}